Refactor simple preprocess event loop to take the place of the old one
