+++ backend/go.mod
@@ -7,7 +7,6 @@ require (
 	github.com/go-logr/logr v1.2.2
 	github.com/golang/mock v1.6.0
 	github.com/google/go-github v17.0.0+incompatible
-	github.com/hashicorp/golang-lru v0.5.1
 	github.com/kcp-dev/logicalcluster/v2 v2.0.0-alpha.3
 	github.com/onsi/ginkgo/v2 v2.1.6
 	github.com/onsi/gomega v1.20.1
